feat : Converts computed latitude and longitude to a float32 and add tests to check the dtype improve the mock datas

+++ satpy/readers/li_l2_nc.py
@@ -158,11 +158,9 @@ class LIL2NCFileHandler(LINCFileHandler):
         data_2d = da.map_blocks(_np_add_at_wrapper, data_2d, (rows, cols), data_array,
                                 dtype=data_array.dtype,
                                 chunks=(LI_GRID_SHAPE[0], LI_GRID_SHAPE[1]))
-        data_2d = da.where(data_2d > 0, data_2d, np.nan)
-
+        data_2d = da.where(data_2d > 0, data_2d, np.float32(np.nan)).astype(np.float32)
         xarr = xr.DataArray(da.asarray(data_2d, CHUNK_SIZE), dims=("y", "x"))
         xarr.attrs = attrs
-
         return xarr
